Move layout components to dedicated directory.
