+++ client/src/screens/AuthScreen.jsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 
 import Carousel from "../components/Carousel";
 import Hero from "../components/Hero";
-import Layout from "../components/Layout";
+import Layout from "../components/Layout/Layout";
 
 export default function AuthScreen() {
     const [input, setInput] = useState("");
