+++ client/src/screens/ProductDetailScreen.jsx
@@ -1,4 +1,4 @@
-import Layout from "../components/Layout";
+import Layout from "../components/Layout/Layout";
 
 export default function ProductDetailScreen() {
     return (
